# storage/api-docs/pos-api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "pDaftar POS Integration API",
        "description": "\n * Integratsiya API — pDaftar POS va uchinchi tomon kassalari (AliPOS, YesPOS va boshqalar) uchun.\n *\n * ## Asosiy qoida\n * pDaftar — **haqiqat manbai** (source of truth). Kassaning lokal bazasi faqat **kesh + navbat**.\n * Qoldiq har doim `stock_movements` ledgeridan hisoblanadi, kassa uni faqat ko'rsatadi.\n *\n * ## Offline ishlash\n * Har bir yozuv amali `client_operation_id` (UUID, kassada amal bajarilgan payt yaratiladi) bilan\n * yuboriladi. Takroriy yuborish hech narsani o'zgartirmaydi — server saqlangan javobni qaytaradi\n * (`replayed: true`). Internet paydo bo'lganda navbat `POST /sync/push` orqali paket qilib yuboriladi.\n *\n * ## Sotuv va qoldiq\n * Kassadan kelgan sotuv **rad etilmaydi**: tovar allaqachon xaridorga berilgan. Qoldiq yetmasa\n * minusga tushadi va bu ko'rinadigan muammo bo'lib qoladi — sinxronlashda yo'qolgan sotuvdan ko'ra yaxshiroq.\n *\n * ## Uchinchi tomon kassalari\n * Tashqi POS tokeni `pos:catalog.read`, `pos:sales.write`, `pos:clients.write` scope'lari bilan\n * beriladi — ular o'z omborini o'zi yuritadi, bizga faqat sotuvni yuboradi.\n * ",
        "contact": {
            "name": "pDaftar",
            "url": "https://pdaftar.uz"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8083",
            "description": "Lokal (Docker)"
        },
        {
...
        }
...
    "paths": {
...
        "/api/pos/v1/sales/recent": {
            "get": {
                "tags": [
                    "POS Sales"
                ],
                "summary": "Do'konning so'nggi sotuvlari",
                "description": "Barcha sotuvchilarning sotuvlari, eng yangisi birinchi. Kassaning o'z navbati emas — bu serverdagi haqiqiy tarix.",
                "operationId": "1d3e855eea63fc6dd7c1d0ea73e1e91b",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "schema": {
                            "type": "integer",
                            "default": 50,
                            "maximum": 200
                        }
                    },
                    {
                        "name": "mine",
                        "in": "query",
                        "description": "1 bo'lsa faqat shu sotuvchining sotuvlari",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "posToken": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "posToken": {
                "type": "http",
                "description": "Kassa tokeni — /terminals/register javobidan olinadi. Ro'yxatdan o'tish uchun oddiy pDaftar user tokeni ishlatiladi.",
                "scheme": "bearer"
            }
        }
    },
...
}